Factored out Model from MMMK
